backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
earthengine-api==0.1.395
google-auth==2.35.0
google-auth-oauthlib==1.2.1
numpy==2.1.2
python-jose[cryptography]==3.3.0
httpx==0.27.0
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
transformers>=4.40.0
rasterio>=1.3.0
albumentations>=1.4.0
